Adequate functionality
When a user connects to the main site a random document id is generated
for them and they are taken to the editor page for the document.

If they share the link to that page with another person, they will both
be connected to the same document and can both edit it. Users are able
to see a current count of how many users are connected to that document.
When all users disconnect from the document its current contents are
saved in the database.

If a user reconnects to a preexisting document its contents are loaded
from the database and the user can resume editing.

__TODOs__
* Make multiple cursors work
  * Currently all users of a document share one cursor
  * Need to have a map from user either pids or user ids to that
    client's cursor
  * Need to handle multiple cursors in the frontend
  * Will I have concurrency issues with multiple people editing at once?
    I want to say no since the document handles incoming edits in serial
    and I currently don't support big modifications like deleting whole
    lines at once
* Handle errors in rendering better
  * Currently if a user tries to use anything that requires a closing
    mark, for example `foo` or __bar__ rendering and possibly updating
    break since I don't handle it will
  * A nice solution would be render everything before the start of the
    broken line as markdown and everything after it as plaintext